% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Funcao para criar o server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Objeto que armazena os inputs da ui}

\item{output}{Objeto que armazena os outputs do server}

\item{session}{Objeto que identifica a sessao do app}
}
\description{
Funcao para criar o server
}
