% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_painel_liquidez.R
\name{body_painel_liquidez}
\alias{body_painel_liquidez}
\title{Funcao que gera o Body Completo do painel usuarios}
\usage{
body_painel_liquidez(con, con2, dt_atualizacao)
}
\arguments{
\item{con}{Conexao com o banco de dados}

\item{con2}{Conexao com o REDFACTOR}
}
\description{
Funcao que gera o Body Completo do painel usuarios
}
