% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{QUERY_DF_INICIAL}
\alias{QUERY_DF_INICIAL}
\title{Query da base de liquidez}
\usage{
QUERY_DF_INICIAL(DATA_REFERENCIA, DT1, DT2, LISTA_OPERADORES)
}
\arguments{
\item{DATA_REFERENCIA}{Data de referencia}

\item{DT1}{Data do Vencimento Inicial}

\item{DT2}{Data do Vencimento Final}

\item{LISTA_OPERADORES}{Lista de Operadores}
}
\description{
Query da base de liquidez
}
