% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{QUERY_DF_TOMADO}
\alias{QUERY_DF_TOMADO}
\title{Query da base de carteira consolidada}
\usage{
QUERY_DF_TOMADO(LISTA_CEDENTE)
}
\arguments{
\item{DATA_REFERENCIA}{Data de referencia}

\item{DT1}{Data do Vencimento Inicial}

\item{DT2}{Data do Vencimento Final}

\item{LISTA_OPERADORES}{Lista de Operadores}
}
\description{
Query da base de carteira consolidada
}
