% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebar}
\alias{sidebar}
\title{Funcao Gerar a sidebar completa do painel usuarios}
\usage{
sidebar(
  con,
  LISTA_TODOS_OS_SUPTS,
  LISTA_TODAS_AS_PLATAFORMAS,
  LISTA_DE_PRODUTOS,
  LISTA_DE_FUNDOS,
  LISTA_CED_APRES_UI,
  FLAG_VIS = 1,
  ALCADA_OP = 1,
  dt_atualizacao,
  dt_vencimento_in
)
}
\arguments{
\item{con}{Conexao com o banco de dados}

\item{LISTA_TODOS_OS_SUPTS}{lista de todos os superintendentes possiveis}

\item{LISTA_TODAS_AS_PLATAFORMAS}{lista de todas os plataformas possiveis}

\item{LISTA_DE_PRODUTOS}{lista de todos os produtos possiveis}

\item{LISTA_DE_FUNDOS}{lista de todos os fundos possiveis}

\item{LISTA_CED_APRES_UI}{Modulo da UI do modelo da apresentação da seleção da escolha da apresentação do cedente}

\item{FLAG_VIS}{Flag to tipo de visualização desejada 1- completo 2- somente capa,3 comente recipocidade, caso contrario pagina em branco}

\item{ALCADA_OP}{Nivel de alçada de visualiação 1- administrativos 2 superintendentes 3 gerentes regioaneis 4 operadores}
}
\description{
Funcao Gerar a sidebar completa do painel usuarios
}
